fix: realocate generic method to header, compiling with the user code
